+++ diesel/src/pg/types/mod.rs
@@ -8,13 +8,16 @@ mod uuid;
 
 #[doc(hidden)]
 pub mod sql_types {
-    #[derive(Debug, Clone, Copy, Default)] pub struct Oid;
     #[derive(Debug, Clone, Copy, Default)] pub struct Array<T>(T);
+    #[derive(Debug, Clone, Copy, Default)] pub struct Oid;
+    #[derive(Debug, Clone, Copy, Default)] pub struct Timestamptz;
+    #[cfg(feature = "uuid")]
+    #[derive(Debug, Clone, Copy, Default)] pub struct Uuid;
+
     pub type SmallSerial = ::types::SmallInt;
     pub type Serial = ::types::Integer;
     pub type BigSerial = ::types::BigInt;
-    #[cfg(feature = "uuid")]
-    #[derive(Debug, Clone, Copy, Default)] pub struct Uuid;
+
     pub type Bytea = ::types::Binary;
     #[doc(hidden)]
     pub type Bpchar = ::types::VarChar;